requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
faiss-cpu
